requirements.txt
Flask==2.3.2
pdfminer.six==20250506
python-docx==0.8.11
pandas==2.2.3
openpyxl==3.1.5
python-magic-bin==0.4.14